+++ src/sdk/PnP.Core/Model/SharePoint/Core/Public/IListItem.cs
@@ -1,4 +1,7 @@
-﻿namespace PnP.Core.Model.SharePoint
+﻿using PnP.Core.Services;
+using System.Threading.Tasks;
+
+namespace PnP.Core.Model.SharePoint
 {
     /// <summary>
     /// Public interface to define a SharePoint Online list item
@@ -20,5 +23,49 @@
         /// Title value of the list item
         /// </summary>
         public string Title { get; set; }
+
+        #region Extension methods
+
+        #region SystemUpdate
+
+        /// <summary>
+        /// Performs a system update of the list item
+        /// </summary>
+        public Task SystemUpdateAsync();
+
+        /// <summary>
+        /// Performs a system update of the list item
+        /// </summary>
+        public void SystemUpdate();
+
+        /// <summary>
+        /// Performs a system update of the list item
+        /// </summary>
+        /// <param name="batch">Batch to add the systemupdate request to</param>
+        public void SystemUpdate(Batch batch);
+
+        #endregion
+
+        #region UpdateOverwriteVersion
+
+        /// <summary>
+        /// Performs a UpdateOverwriteVersion of the list item
+        /// </summary>
+        public Task UpdateOverwriteVersionAsync();
+
+        /// <summary>
+        /// Performs a UpdateOverwriteVersion of the list item
+        /// </summary>
+        public void UpdateOverwriteVersion();
+
+        /// <summary>
+        /// Performs a UpdateOverwriteVersion of the list item
+        /// </summary>
+        /// <param name="batch">Batch to add the UpdateOverwriteVersion request to</param>
+        public void UpdateOverwriteVersion(Batch batch);
+
+        #endregion
+
+        #endregion
     }
 }